// types/party.ts
export interface PartyFormData {
  PartyId: number;
  PartyName: string;
  PartyNumber: string;
  PartyType: string;
  ParyBalaceType: string;
  PartyBalaceAmt: string;
  IsGST: boolean;
  GSTNumber: string;
  PAN: string;
  IsBillingAdd: boolean;
  BillingAdd: string;
  PINcode: string;
  BillingState: string;
  IsShippingSame: boolean;
  IsBankAccount: boolean;
  AccountNumber: string;
  HolderName: string;
  IFSCCode: string;
  BankName: string;
  Branch: string;
  CompnayID: number;
  StoreId: number;
  UserID: number;
}